+++ apps/block_scout_web/lib/block_scout_web/web_router.ex
@@ -178,6 +178,20 @@ defmodule BlockScoutWeb.WebRouter do
         only: [:index],
         as: :inventory
       )
+
+      resources(
+        "/instance",
+        Tokens.InstanceController,
+        only: [:show],
+        as: :instance
+      ) do
+        resources(
+          "/token_transfers",
+          Tokens.Instance.TransferController,
+          only: [:index],
+          as: :transfer
+        )
+      end
     end
 
     resources(